+++ src/bucketconfig/bc_cccp.c
@@ -442,6 +442,34 @@ static void request_config(cccp_provider *cccp)
     lcbio_timer_rearm(cccp->timer, PROVIDER_SETTING(&cccp->base, config_node_timeout));
 }
 
+static void do_dump(clconfig_provider *pb, FILE *fp)
+{
+    cccp_provider *cccp = (cccp_provider *)pb;
+    unsigned ii;
+
+    if (!cccp->base.enabled) {
+        return;
+    }
+
+    fprintf(fp, "## BEGIN CCCP PROVIDER DUMP ##\n");
+    fprintf(fp, "TIMER ACTIVE: %s\n", lcbio_timer_armed(cccp->timer) ? "YES" : "NO");
+    fprintf(fp, "PIPELINE RESPONSE COOKIE: %p\n", cccp->cmdcookie);
+    if (cccp->ioctx) {
+        fprintf(fp, "CCCP Owns connection:\n");
+        lcbio_ctx_dump(cccp->ioctx, fp);
+    } else if (cccp->creq.u.p_generic) {
+        fprintf(fp, "CCCP Is connecting\n");
+    } else {
+        fprintf(fp, "CCCP does not have a dedicated connection\n");
+    }
+
+    for (ii = 0; ii < cccp->nodes->nentries; ii++) {
+        const lcb_host_t *curhost = &cccp->nodes->entries[ii];
+        fprintf(fp, "CCCP NODE: %s:%s\n", curhost->host, curhost->port);
+    }
+    fprintf(fp, "## END CCCP PROVIDER DUMP ##\n");
+}
+
 clconfig_provider * lcb_clconfig_create_cccp(lcb_confmon *mon)
 {
     cccp_provider *cccp = calloc(1, sizeof(*cccp));
@@ -454,6 +482,7 @@ clconfig_provider * lcb_clconfig_create_cccp(lcb_confmon *mon)
     cccp->base.config_updated = config_updated;
     cccp->base.configure_nodes = configure_nodes;
     cccp->base.get_nodes = get_nodes;
+    cccp->base.dump = do_dump;
     cccp->base.parent = mon;
     cccp->base.enabled = 0;
     cccp->timer = lcbio_timer_new(mon->iot, cccp, socket_timeout);