+++ src/bucketconfig/confmon.c
@@ -50,6 +50,7 @@ provider_string(clconfig_method_t type) {
     if (type == LCB_CLCONFIG_CCCP) { return "CCCP"; }
     if (type == LCB_CLCONFIG_FILE) { return "FILE"; }
     if (type == LCB_CLCONFIG_MCRAW) { return "MCRAW"; }
+    if (type == LCB_CLCONFIG_USER) { return "USER"; }
     return "";
 }
 
@@ -418,4 +419,39 @@ lcb_confmon_set_provider_active(lcb_confmon *mon,
         provider->enabled = enabled;
     }
     lcb_confmon_prepare(mon);
+}
+
+void
+lcb_confmon_dump(lcb_confmon *mon, FILE *fp)
+{
+    unsigned ii;
+    fprintf(fp, "CONFMON=%p\n", (void*)mon);
+    fprintf(fp, "STATE= (0x%x)", mon->state);
+    if (mon->state & CONFMON_S_ACTIVE) {
+        fprintf(fp, "ACTIVE|");
+    }
+    if (mon->state == CONFMON_S_INACTIVE) {
+        fprintf(fp, "INACTIVE/IDLE");
+    }
+    if (mon->state & CONFMON_S_ITERGRACE) {
+        fprintf(fp, "ITERGRACE");
+    }
+    fprintf(fp, "\n");
+    fprintf(fp, "LAST ERROR: 0x%x\n", mon->last_error);
+
+
+    for (ii = 0; ii < LCB_CLCONFIG_MAX; ii++) {
+        clconfig_provider *cur = mon->all_providers[ii];
+        if (!cur) {
+            continue;
+        }
+
+        fprintf(fp, "** PROVIDER: 0x%x (%s) %p\n", cur->type, provider_string(cur->type), cur);
+        fprintf(fp, "** ENABLED: %s\n", cur->enabled ? "YES" : "NO");
+        fprintf(fp, "** CURRENT: %s\n", cur == mon->cur_provider ? "YES" : "NO");
+        if (cur->dump) {
+            cur->dump(cur, fp);
+        }
+        fprintf(fp, "\n");
+    }
 }