+++ tools/common/options.h
@@ -23,7 +23,8 @@ namespace cbc {
     X(String, capath, "capath", '\0') \
     X(UInt, timeout, "timeout", 't') \
     X(Bool, timings, "timings", 'T') \
-    X(Bool, verbose, "verbose", 'v')
+    X(Bool, verbose, "verbose", 'v') \
+    X(Bool, dump, "dump", '\0')
 
 
 class ConnParams {
@@ -34,6 +35,7 @@ public:
     lcb_error_t doCtls(lcb_t instance);
     bool useTimings() { return o_timings.result(); }
     void setAdminMode();
+    bool shouldDump() { return o_dump.result(); }
     void writeConfig(const std::string& dest = getConfigfileName());
     static std::string getConfigfileName();